ingestion/requirements.txt
aiobotocore==2.14.0
aiohappyeyeballs==2.3.6
aiohttp==3.10.3
aioitertools==0.12.0
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
asgiref==3.8.1
async-timeout==4.0.3
asyncpg==0.29.0
attrs==24.2.0
beautifulsoup4==4.12.3
bentoml==1.3.2
boto3==1.35.0
botocore==1.35.0
cattrs==23.1.2
certifi==2024.7.4
charset-normalizer==3.3.2
circus==0.18.0
click==8.1.7
click-option-group==0.5.6
cloudpickle==3.0.0
dataclasses-json==0.6.7
deepmerge==1.1.1
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
docx2txt==0.8
dulwich==0.22.1
et-xmlfile==1.1.0
exceptiongroup==1.2.2
filelock==3.15.4
fpdf==1.7.2
frozenlist==1.4.1
fs==2.4.16
fsspec==2024.6.1
greenlet==3.0.3
h11==0.14.0
html2text==2024.2.26
httpcore==1.0.5
httpx==0.27.0
httpx-ws==0.6.0
huggingface-hub==0.24.5
idna==3.7
importlib-metadata==6.11.0
inflection==0.5.1
inquirerpy==0.3.4
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.14
langchain-community==0.2.12
langchain-core==0.2.32
langchain-text-splitters==0.2.2
langsmith==0.1.99
llama-cloud==0.0.13
llama-index==0.10.65
llama-index-agent-openai==0.2.9
llama-index-cli==0.1.13
llama-index-core==0.10.66
llama-index-embeddings-openai==0.1.11
llama-index-embeddings-text-embeddings-inference==0.1.4
llama-index-indices-managed-llama-cloud==0.2.7
llama-index-legacy==0.9.48.post2
llama-index-llms-openai==0.1.27
llama-index-llms-openai-like==0.1.3
llama-index-llms-openllm==0.2.0
llama-index-multi-modal-llms-openai==0.1.9
llama-index-program-openai==0.1.7
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.33
llama-index-readers-llama-parse==0.1.6
llama-index-readers-s3==0.1.10
llama-index-utils-huggingface==0.1.1
llama-index-vector-stores-postgres==0.1.14
llama-parse==0.4.9
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mdurl==0.1.2
minijinja==2.0.1
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
nvidia-cublas-cu12==12.3.4.1
nvidia-cuda-cupti-cu12==12.3.101
nvidia-cuda-nvrtc-cu12==12.3.107
nvidia-cuda-runtime-cu12==12.3.101
nvidia-cudnn-cu12==9.1.1.17
nvidia-cufft-cu12==11.0.12.1
nvidia-curand-cu12==10.3.4.107
nvidia-cusolver-cu12==11.5.4.101
nvidia-cusparse-cu12==12.2.0.103
nvidia-ml-py==11.525.150
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.3.101
openai==1.38.0
openllm==0.6.9
openpyxl==3.1.5
opentelemetry-api==1.20.0
opentelemetry-instrumentation==0.41b0
opentelemetry-instrumentation-aiohttp-client==0.41b0
opentelemetry-instrumentation-asgi==0.41b0
opentelemetry-sdk==1.20.0
opentelemetry-semantic-conventions==0.41b0
opentelemetry-util-http==0.41b0
orjson==3.10.7
packaging==24.1
pandas==2.2.2
pathlib==1.0.1
pathspec==0.12.1
pfzy==0.3.4
pgvector==0.2.5
pillow==10.4.0
pip-requirements-parser==32.0.1
prometheus_client==0.20.0
prompt-toolkit==3.0.36
psutil==6.0.0
psycopg2-binary==2.9.9
pyaml==24.7.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyparsing==3.1.2
pypdf==4.3.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.9
python-pptx==1.0.2
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.1.0
questionary==2.0.1
regex==2024.7.24
requests==2.32.3
rich==13.7.1
s3fs==2024.6.1
s3transfer==0.10.2
safetensors==0.4.4
schema==0.7.7
shellingham==1.5.4
simple-di==0.1.5
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.32
starlette==0.38.2
striprtf==0.0.26
sympy==1.13.3
tabulate==0.9.0
tenacity==8.5.0
tiktoken==0.7.0
tokenizers==0.19.1
tomli==2.0.1
tomli_w==1.0.0
torch==2.4.1
tornado==6.4.1
tqdm==4.66.5
transformers==4.44.0
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uv==0.2.37
uvicorn==0.30.6
watchfiles==0.23.0
wcwidth==0.2.13
wrapt==1.16.0
wsproto==1.2.0
XlsxWriter==3.2.0
yarl==1.9.4
zipp==3.20.0